Pass error message through
The existing error message suggests that ananas is having trouble loading the bot module, which can cause some troubleshooting difficulty when it's actually a module the *bot itself* is calling that's causing the ModuleNotFoundError. This change reports the name of the missing module to the user, so that they can more easily solve the problem.